[#58805] Skift felttype for CPR så alle tal kan stå der

See merge request gronlandsprojekter/toldbehandling!128

+++ rest/common/models.py
@@ -12,7 +12,7 @@ class IndberetterProfile(models.Model):
     user = models.OneToOneField(
         User, on_delete=models.CASCADE, null=False, related_name="indberetter_data"
     )
-    cpr = models.PositiveIntegerField(
+    cpr = models.BigIntegerField(
         null=True,
         blank=True,
         validators=[
@@ -50,7 +50,7 @@ class Postnummer(models.Model):
 
 class EboksBesked(models.Model):
     titel = models.CharField(max_length=500)
-    cpr = models.PositiveIntegerField(
+    cpr = models.BigIntegerField(
         null=True,
         blank=True,
         validators=[
